+++ cartoframes/carto_vl/__init__.py
@@ -1,18 +1,14 @@
 from __future__ import absolute_import
 
-from .basemap.basemaps import Basemaps as basemaps
-from .utils import defaults
-from .utils.html import HTMLMap
-from .maps.map import Map
-from .layer.layer import Layer
-from .style.style import Style
+from .basemaps import Basemaps as basemaps
+from .map import Map
+from .layer import Layer
+from .style import Style
 from .source import Source
 
 
 __all__ = [
-    'defaults',
     'basemaps',
-    'HTMLMap',
     'Map',
     'Layer',
     'Source',
